cleaned up process launch mode handling significantly
Now the delegate is the only one to define how to launch a process. Now it's possible to spawn, replace, or fork
a process, and all of the modes have a real-work requirement and need.

Also the interface of the process controller is not cluttered anymore with a special need, and
launch_mode replaced the previous spawn flag.

In short, it's much better now, and allows to write cleaner delegates