This fixes a few issues with global state maintained incorrectly during test-cases.
There is just one spot in bcore where this is done, and it's already a source for trouble.
The issue is rooted in the necessity to derive from plugins to auto-register their types.
Original source of the issue is the reliance of tests to use some plugins which are standard, and are always loaded.
Maybe this is the issue - plugins should never be always loaded, or just not be a plugin