Implemented quick and dirty cache to reduce yaml loading times
The cache writes to a local temp directory, and reads from it as well.
It must use pickle, unfortunately, which isn't the fastests, but better than anything else that would work.
It should of course be possible to disable the cache, even though in common operation you would always want it.
It won't reduce the IOps at all (especially those over the network), but merely reduces yaml decode time.