Fixed type bugs introduced previously.
Just did a spike which was based on false assumptions, as I thought saving some MD5 operations through a cache
could save 30ms. It was too good to be true, anyway.

Note that test-cases are not truly self-contained anymore as they will re-read their caches after the first run or testcase.
This must be configurable, but this will be done latest when runtime configuration reloading is implemented.